synapta_src/backend/requirements.txt
mlx_lm
sentence-transformers
fastapi
uvicorn
pydantic
numpy
safetensors
torch
transformers
peft
accelerate
